\name{sphereplot-package}
\alias{sphereplot-package}
\alias{sphereplot}
\docType{package}
\title{
Spherical plotting routines
}
\description{
Various low level and high level routines for generate spherical plots. Includes celestial sphere style 3D grid and routines for overlaying additional points and text. Requires rgl.
}
\details{
\tabular{ll}{
Package: \tab sphereplot\cr
Type: \tab Package\cr
Version: \tab 1.5\cr
Date: \tab 2013-09-22\cr
License: \tab GPL-2\cr
}
Standard usage is to run rgl.sphgrid to make the 3D coordinate grid, then add points using rgl.sphpoints.
}
\author{
Aaron Robotham

Maintainer: Aaron Robotham <aaron.robotham@uwa.edu.au>
}
\keyword{ package }

